melhoria do microsserviço de cliente

+++ mscliente/src/main/java/com/rai/mscliente/repository/ClienteRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ClienteRepository extends JpaRepository<Cliente, Integer> {
+public interface ClienteRepository extends JpaRepository<Cliente, Long> {
 
     Optional<Cliente> findClienteByCpf(String cpf);
 }
